ScreenSaver-V1.0
This is the first attempt at creating screensavers with SFML. This version is horribly inefficient, and will be updated soon.
This version was created to look and act like the old DVD player screensavers, however with some RGB. Due to an oversight, some colour combinations are not present, which will be fixed in a later update.
Finally, this has 2 different variants. One is a dark mode, while the other is not. Choose wisely.